update logic for sending messages via notify
