+++ src/messaging/inbound/comms-request/handler.js
@@ -4,8 +4,6 @@ import { getCommsProcessor } from './processors/processor.js'
 
 import { parseSqsMessage } from '../../sqs/parse-message.js'
 
-import { sendNotification } from './send-notification.js'
-
 const logger = createLogger()
 
 const handleCommRequestMessages = async (messages) => {
@@ -18,8 +16,6 @@ const handleCommRequestMessages = async (messages) => {
       const processor = getCommsProcessor(content)
 
       await processor(content)
-
-      await sendNotification(message)
     } catch (err) {
       logger.error(`Error processing message: ${err.message}`)
     }
