+++ src/messaging/inbound/comms-request/processors/v3.js
@@ -5,6 +5,8 @@ import { v3 } from '../../../../schemas/comms-request/index.js'
 
 import { checkNotificationIdempotency, addNotificationRequest } from '../../../../repos/notification-log.js'
 
+import { trySendViaNotify } from '../try-send-via-notify.js'
+
 const logger = createLogger()
 
 const processV3CommsRequest = async (message) => {
@@ -19,6 +21,14 @@ const processV3CommsRequest = async (message) => {
   }
 
   await addNotificationRequest(validated)
+
+  const emailAddresses = Array.isArray(message.data.commsAddresses)
+    ? message.data.commsAddresses
+    : [message.data.commsAddresses]
+
+  for (const emailAddress of emailAddresses) {
+    await trySendViaNotify(message, emailAddress)
+  }
 
   return logger.info(`Comms V3 request processed successfully, eventId: ${validated.id}`)
 }
